backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
supabase==2.7.4
python-dotenv==1.0.1
